+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SpaghettiChef">
+        android:theme="@style/Theme.SpaghettiChef"
+        android:usesCleartextTraffic="true">
         <activity
             android:name=".view.MainActivity"
             android:exported="true">
