+++ app/src/main/java/org/sopt/spaghettichef/presenter/TodoPresenter.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import androidx.lifecycle.LifecycleCoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import org.sopt.spaghettichef.model.TodoDatabase
-import org.sopt.spaghettichef.model.TodoItem
+import org.sopt.spaghettichef.model.db.todo.TodoDatabase
+import org.sopt.spaghettichef.model.db.todo.TodoItem
 
 class TodoPresenter(
     private val view: TodoContract.View,
